(PC-20298)[PRO] fix: rename SignupOfferer to SignupJourney
