+++ pro/src/components/SignupJourneyBreadcrumb/index.ts
@@ -1,0 +1,1 @@
+export { default as SignupJourneyBreadcrumb } from './SignupJourneyBreadcrumb'
