+++ pro/src/components/SignupJourneyFormLayout/index.ts
@@ -1,0 +1,1 @@
+export { default as SignupJourneyFormLayout } from './SignupJourneyFormLayout'
