--- pro/src/components/SignupOffererFormLayout/index.ts
@@ -1,1 +1,0 @@
-export { default as SignupOffererFormLayout } from './SignupOffererLayout'
