+++ pro/src/pages/SignupOffererRoutes/index.ts
@@ -1,0 +1,1 @@
+export { default as SignupJourneyRoutes } from './SignupJourneyRoutes'
