+++ pro/src/screens/SignupJourneyForm/types.ts
@@ -1,0 +1,3 @@
+import { IActivityFormValues } from './Activity/ActivityForm'
+
+export type ISignupJourneyFormValues = IActivityFormValues
